Regexp: save the original sed command arguments
So we can have access to the original arguments, as informed by
the user. This is useful since self.args gets overwritten with
processes values (Python-flavored regexes, & turns to \g<0>).

XXX
So when dumping with -d the original sed regex will be shown
in the s command. Currently it shows the Python-converted regex.